moving all utils/keymaps modules to a single folder
