+++ app/utils/keymaps/find-command-by-keys.js
@@ -1,4 +1,4 @@
-const normalize = require('./keymaps-normalize');
+const normalize = require('./normalize');
 
 module.exports = (keys, commands) => {
   return commands[normalize(keys)];
