+++ app/utils/keymaps/get-command.js
@@ -1,0 +1,6 @@
+const {getKeymaps} = require('../../config');
+const findCommandByKeys = require('./find-command-by-keys');
+
+module.exports = keys => {
+  return findCommandByKeys(keys, getKeymaps().keys);
+};
